requirements.txt
aiosignal==1.2.0
alabaster==0.7.12
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
binaryornot==0.4.4
black==22.3.0
boolean.py==3.8
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.7.0
Django==4.0.7
django-environ==0.8.1
django-filter==21.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
doc8==0.11.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.17.1
executing==0.8.3
freezegun==1.2.1
frozenlist==1.3.0
gitdb==4.0.9
GitPython==3.1.27
gunicorn==20.1.0
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.3
iniconfig==1.1.1
ipython==8.0.1
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.1
jsonschema==3.2.0
license-expression==21.6.14
lxml==4.9.1
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
multidict==6.0.2
mypy-extensions==0.4.3
packageurl-python==0.9.9
packaging==21.3
paramiko==2.10.3
parso==0.8.3
pathspec==0.9.0
pbr==5.8.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.1
pluggy==1.0.0
pprintpp==0.4.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyNaCl==1.5.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.1.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==5.4.1
requests==2.27.1
restructuredtext-lint==1.4.0
saneyaml==0.5.2
semantic-version==2.9.0
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
soupsieve==2.3.2
Sphinx==4.5.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-django2==1.5
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
stack-data==0.2.0
stevedore==3.5.0
texttable==1.6.4
toml==0.10.2
tomli==2.0.1
traitlets==5.1.1
typing_extensions==4.1.1
univers==30.7.0
urllib3==1.26.9
wcwidth==0.2.5
websocket-client==0.59.0
yarl==1.7.2
zipp==3.8.0
dateparser==1.1.1
fetchcode==0.1.0
